Extract logic for "Unload client" button out of `serve-dev.js`
Per discussion on the PR, it is preferable to put markup and JS in the
dev server in external files rather than as string literals in
`serve-dev.js`. This enables them to be edited without having to restart
the dev server.

 - Extract shared functions for loading and unloading client into
   `scripts/util.js`

 - Move logic specific to the index template into `scripts/index.js`.
