+++ dev-server/static/scripts/index.js
@@ -1,0 +1,16 @@
+import { activeClientUrl, loadClient, unloadClient } from './util.js';
+
+const toggleClientButton = document.querySelector('.js-toggle-client');
+toggleClientButton.textContent = 'Unload client';
+const clientUrl = activeClientUrl;
+
+toggleClientButton.onclick = () => {
+  if (activeClientUrl) {
+    unloadClient();
+    toggleClientButton.textContent = 'Load client';
+  } else {
+    loadClient(clientUrl);
+    toggleClientButton.textContent = 'Unload client';
+  }
+};
+
