+++ dev-server/templates/index.mustache
@@ -49,5 +49,6 @@
     <li><a href="https://h.readthedocs.io/en/latest/api/">Hypothesis API documentation</a></li>
   </ul>
   {{{hypothesisScript}}}
+  <script type="module" src="/scripts/index.js"></script>
 </body>
 </html>
